move common rule stuff to clangcommon
